### 2016-07-28 #46 @whedon
[REVIEW]: pygtc: beautiful parameter covariance plots (aka. Giant Triangle Confusograms)
**Submitting author:** @SebastianBocquet (<a href="http://orcid.org/0000-0002-4900-805X">Sebastian Bocquet</a>)
**Repository:** <a href="https://github.com/SebastianBocquet/pygtc" target ="_blank">https://github.com/SebastianBocquet/pygtc</a>
**Version:** 0.1.1
**Archive:** <a href="https://dx.doi.org/10.5281/zenodo.159225" target ="_blank">https://dx.doi.org/10.5281/zenodo.159225</a>
