### 2016-08-01 #46 @arfon
@dfm ✨

I'll need to add you to the `@openjournals/joss-reviewers` team so you have the correct permissions to update the checklist at the top of the issue - sound OK?
