### 2016-08-01 #46 @dfm
### General checks

> Version: Does the release version given match the GitHub release (0.1.1)?

There is no archived release linked from the documentation or readme and there is no DOI given in the paper.

### Functionality

> Installation: Does installation proceed as outlined in the documentation?

The installation instructions are missing a list of prerequisites. Does scipy really need to be a hard dependency?

### Documentation

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

There is a vague statement of need in the README but it's not very clear and it is completely missing from the docs. The contents of the paper (and maybe more discussion) should be included in the docs. There is also no discussion of how to interpret these plots. For example, the (incorrect, [I would argue](http://corner.readthedocs.io/en/latest/pages/sigmas.html)!) choice of contour levels should be stated and justified.

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

In principle, the pip command could install all of the dependencies but they should be discussed in the installation instructions.

> Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

There is a demo notebook included in the repository but the docs have no discussion of the usage or even a link to the demo. I think that a gallery of examples or at least a tutorial is required.

> Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g. API method documentation)?

The entry point is documented by its docstring and an auto-generated docs entry. This isn't really sufficient to constitute documentation.

> Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

No tests are currently included. It might be worth checking out the tests from [corner.py](https://github.com/dfm/corner.py) for an example of how this could be implemented for similar software.

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Missing. There should (at least!) be a link to the GitHub repository from the docs.

### Software paper

> A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

This is well described but the paper incorrectly states that corner.py can't overplot multiple samples. This has never been true! 😉 The claim is also made that the resulting plots are compatible with different journals and a list of options are given but it's not clear to me why this is necessary and why the dpi enters - no one should be submitting rasterized figures to the journals!

> References: Do all archival references that should have a DOI list one (e.g. papers, datasets, software)?

The DOI for numpy/scipy is 10.1109/MCSE.2011.37 and matplotlib is 10.1109/MCSE.2007.55. [ref](https://www.scipy.org/citing.html)

### Other comments

Packages should **never** alter the matplotlib rc. The fact that plots made _after_ calling pygtc are different from the same plots made before is not ok.
