+++ docs/index.rst
@@ -3,9 +3,11 @@ pygtc.py
 Make a publication-ready giant triangle confusogram (GTC, aka triangle/corner
 plot) with just one line of code.
 
+The pygtc module uses matplotlib to display samples of multi-dimensional likelihood functions produced by MCMC or similar algorithms in a beautiful figure.
+
 Installation
 ============
-Installation is as easy as ``pip install pygtc``
+pygtc depends on ``numpy``, ``matplotlib``, and ``scipy``. Installation is as easy as ``pip install pygtc``.
 
 API
 ===
