+++ docs/index.rst
@@ -1,13 +1,24 @@
 pygtc.py
---------
-Make a publication-ready giant triangle confusogram (GTC, aka triangle/corner
-plot) with just one line of code.
+===
+Make a publication-ready giant triangle confusogram with just one line of code.
 
-The pygtc module uses matplotlib to display samples of multi-dimensional likelihood functions produced by MCMC or similar algorithms in a beautiful figure.
+
+What is a Giant Triangle Confusogram?
+    A Giant Triangle Confusogram (GTC, aka triangle/corner plot) is a way of displaying the results of an MCMC
+    or similar analysis. The recovered parameter constraints are displayed on a grid in which the diagonal shows
+    the one-dimensional posteriors and the lower-left triangle shows the pairwise projections.
+
 
 Installation
-============
+---
 pygtc depends on ``numpy``, ``matplotlib``, and ``scipy``. Installation is as easy as ``pip install pygtc``.
+
+
+Example Usage
+---
+We recommend to familiarize yourself with pygtc's main features by studying our tutorial
+`demo.ipynp <https://github.com/SebastianBocquet/pygtc/blob/master/demo.ipynb>`_
+
 
 API
 ===
