+++ paper/paper.bib
@@ -10,7 +10,8 @@
   publication-quality image generation across user
   interfaces and operating systems.},
   publisher = {IEEE COMPUTER SOC},
-  year      = 2007
+  year      = {2007},
+  doi = {10.1109/MCSE.2007.55}
 }
 
 @article{emcee,
@@ -82,5 +83,6 @@ archivePrefix = "arXiv",
   title =     {{SciPy}: Open source scientific tools for {Python}},
   year =      {2001--},
   url = "http://www.scipy.org/",
-  note = {[Online; accessed 2016-07-27]}
+  note = {[Online; accessed 2016-07-27]},
+  doi = {10.1109/MCSE.2011.37}
 }
