### 2016-08-01 39b4dc3
removed wrong statement about corner.py
