### 2016-08-02 399ec79
set font sizes directly instead of modifying rcParams; scipy as option
