+++ pygtc/pygtc.py
@@ -1,9 +1,13 @@
 from matplotlib import pyplot as plt
 import numpy as np
-import scipy.ndimage
 import matplotlib.ticker as mtik
-from scipy.stats import norm
-
+try:
+    import scipy.ndimage
+    from scipy.stats import norm
+    haveScipy = True
+except ImportError:
+    haveScipy = False
+    
 __all__ = ['plotGTC']
 
 
@@ -81,9 +85,9 @@ def plotGTC(chains, **kwargs):
         An integer describing the number of bins used to compute the
         histograms. Default is 30.
 
-    smoothingKernel : int
-        An integer describing the size of the Gaussian smoothing kernel in
-        bins. Default is 1.
+    smoothingKernel : float
+        A number describing the size of the Gaussian smoothing kernel in
+        bins. Default is 1. Set to 0 for no smoothing.
 
     figureSize : float or string
         A number in inches describing the length = width of the GTC, or a
@@ -119,8 +123,12 @@ def plotGTC(chains, **kwargs):
         is ``True``.
 
     doOnly1dPlot : bool
-        Pot only ONE 1d histogram. If this is True, then chains must have
+        Plot only ONE 1d histogram. If this is True, then chains must have
         shape ``(samples,1)``. Default is ``False``.
+    
+    mathTextFontSet : string
+        Set consistent font family in rcParams. Default is ``stixsans``. Set
+        to ``'None'`` to use the setting in your matplotlib rc.
 
     Returns
     -------
@@ -132,18 +140,7 @@ def plotGTC(chains, **kwargs):
 
     """
 
-    ##### Matplotlib and figure setting
-    # Matplotlb rcParams
-    axisColor = '#333333'
-    plt.rcParams['legend.fontsize'] = 9
-    plt.rcParams['axes.labelsize'] = 9
-    plt.rcParams['xtick.labelsize'] = 6
-    plt.rcParams['ytick.labelsize'] = 6
-    plt.rcParams['axes.edgecolor'] = axisColor
-    plt.rcParams['xtick.color'] = axisColor
-    plt.rcParams['ytick.color'] = axisColor
-    plt.rcParams['text.usetex'] = True
-    plt.rcParams['text.latex.preamble'] = [r'\usepackage{sansmath}', r'\sansmath']
+    ##### Figure setting
 
     #Set up some colors
     truthsDefaultColors = ['#4d4d4d', '#4d4d4d', '#4d4d4d']
@@ -276,6 +273,7 @@ def plotGTC(chains, **kwargs):
     # Show Gaussian PDF on 1d plots (to show Gaussian priors)
     priors = kwargs.pop('priors', None)
     if priors is not None:
+        assert haveScipy==True, "You need to have scipy installed to display Gaussian priors"
         assert len(priors)==nDim, "List of priors must match number of parameters"
         for i in range(nDim):
             if priors[i]:
@@ -306,6 +304,8 @@ def plotGTC(chains, **kwargs):
     smoothingKernel = kwargs.pop('smoothingKernel', 1) #Don't you like smooth data?
     if smoothingKernel>=nBins/10:
         print("Wow, that's a huge smoothing kernel! You sure you want its scale to be %.1f percent of the plot?!"%(100.*float(smoothingKernel)/float(nBins)))
+    if not haveScipy:
+        smoothingKernel = 0
 
     # Figure size: choose size to fit journal, use reasonable default, or provide your own
     figureSize = kwargs.pop('figureSize', None) #Figure size descriptor or figure width=height in inches
@@ -345,6 +345,11 @@ def plotGTC(chains, **kwargs):
[…]
