### 2016-08-03 497f920
docs: scipy is only optional
