### 2016-08-03 79cbc3c
paper.md : scipy is optional
