### 2016-08-03 19592c2
Hardcode ax bgcolor and no grid
