### 2016-08-03 465d742
hardcode line widths
