### 2016-08-04 7d49c6a
Many hardcoded plot style options
