+++ pygtc/pygtc.py
@@ -348,7 +348,7 @@ def plotGTC(chains, **kwargs):
     # Set font in rcParams
     mathTextFontSet = kwargs.pop('mathTextFontSet', 'stixsans')
     if not mathTextFontSet=='None':
-        plt.rcParams['mathtext.fontset'] = mathTextFontSet
+        plt.rcParams['mathtext.fontset'] = mathTextFontSet    
         
     # Check to see if there are any remaining keyword arguments
     keys = ''
@@ -368,7 +368,7 @@ def plotGTC(chains, **kwargs):
     axisColor = '#333333'
     # Create the figure, and empty list for first column / last row
     fig = plt.figure(figsize=(figureWidth,figureWidth))
-    axV, axH = [],[]
+    axV, axH = [], []
 
 
 
@@ -382,11 +382,6 @@ def plotGTC(chains, **kwargs):
                         ax = fig.add_subplot(nDim,nDim,(i*nDim)+j+1)
                     else:
                         ax = fig.add_subplot(nDim-1,nDim-1,((i-1)*(nDim-1))+j+1)
-                    
-                    # Layout things
-                    ax.set_axis_bgcolor('w')
-                    ax.tick_params(direction='in')
-                    ax.grid(False)
                     
                     ##### Draw contours and truths
                     # Extract 2d chains
@@ -425,6 +420,16 @@ def plotGTC(chains, **kwargs):
                     else:
                         ax.get_yaxis().set_ticklabels([])
 
+                    ##### Panel layout
+                    ax.grid(False)
+                    #ax.set_axis_bgcolor('none')
+                    for axis in ['top','bottom','left','right']:
+                        ax.spines[axis].set_color(axisColor)
+                        ax.spines[axis].set_linewidth(1)
+                        
+                    ##### Global tick properties
+                    ax.tick_params(direction='in', pad=4, colors=axisColor, size=4, width=.5, labelsize=6)
+                    
                     ##### Rotate tick labels
                     for xLabel in ax.get_xticklabels():
                         xLabel.set_rotation(tickAngle)
@@ -460,22 +465,11 @@ def plotGTC(chains, **kwargs):
                     ax.yaxis.set_ticks(yTicks[i])
                                         
                     ##### First column and last row are needed to align labels
-                    # Set label font size
                     if j==0:
                         axV.append(ax)
-                        ax.tick_params(labelsize=6)
-                        
                     if i==nDim-1:
                         axH.append(ax)
-                        ax.tick_params(labelsize=6)
-                    
-                    ##### Light black axes 
-                    ax.spines['bottom'].set_color(axisColor)
-                    ax.spines['top'].set_color(axisColor) 
-                    ax.spines['right'].set_color(axisColor)
-                    ax.spines['left'].set_color(axisColor)
-                    ax.tick_params(axis='x', colors=axisColor)
-                    ax.tick_params(axis='y', colors=axisColor)
+                                        
                         
                     
 
@@ -484,11 +478,6 @@ def plotGTC(chains, **kwargs):
         for i in range(nDim):
             ##### Create subplot
             ax = fig.add_subplot(nDim,nDim,(i*nDim)+i+1)
-
-            # Layout things
-            ax.set_axis_bgcolor('w')
-            ax.tick_params(direction='in')
-            ax.grid(False)
                                 
             ##### Plot histograms, truths, Gaussians
             # Extract 1d chains
@@ -506,6 +495,16 @@ def plotGTC(chains, **kwargs):
[…]
