### 2016-08-05 4ca5426
new features: point density; no filled contours
