+++ pygtc/pygtc.py
@@ -1,4 +1,7 @@
 from matplotlib import pyplot as plt
+from matplotlib.colors import LinearSegmentedColormap
+#from matplotlib.colors import LogNorm
+import matplotlib.colors as mplcolors
 import numpy as np
 import matplotlib.ticker as mtik
 try:
@@ -89,6 +92,13 @@ def plotGTC(chains, **kwargs):
         A number describing the size of the Gaussian smoothing kernel in
         bins. Default is 1. Set to 0 for no smoothing.
 
+    filledPlots : bool
+        Whether you want the 2d contours and the 1d histograms to be 
+        filled. Default is ``True``.
+    
+    plotDensity : bool
+        Whether you want to see the 2d density of points. Default is ``False``.
+    
     figureSize : float or string
         A number in inches describing the length = width of the GTC, or a
         string indicating a predefined journal setting and whether the
@@ -302,11 +312,18 @@ def plotGTC(chains, **kwargs):
     # Data binning and smoothing
     nBins = kwargs.pop('nBins', 30) # Number of bins for 1d and 2d histograms. 30 works...
     smoothingKernel = kwargs.pop('smoothingKernel', 1) #Don't you like smooth data?
+    if not haveScipy:
+        print "Warning: You don't have Scipy installed. Your curves will not be smoothed."
+        smoothingKernel = 0
     if smoothingKernel>=nBins/10:
         print("Wow, that's a huge smoothing kernel! You sure you want its scale to be %.1f percent of the plot?!"%(100.*float(smoothingKernel)/float(nBins)))
-    if not haveScipy:
-        smoothingKernel = 0
 
+    # Filled contours and histograms
+    filledPlots = kwargs.pop('filledPlots', True)
+    
+    # Filled contours and histograms
+    plotDensity = kwargs.pop('plotDensity', False)
+    
     # Figure size: choose size to fit journal, use reasonable default, or provide your own
     figureSize = kwargs.pop('figureSize', None) #Figure size descriptor or figure width=height in inches
     if figureSize is None:
@@ -356,13 +373,10 @@ def plotGTC(chains, **kwargs):
         keys = keys + key + ' '
         raise NameError("illegal keyword arguments: " + keys)
 
-    # These are needed to compute the confidence levels
-    nBinsFlat = np.linspace(0., nBins**2, nBins**2)
+    
 
-    # Left and right panel boundaries
-    panelXrange = np.empty((nDim,2))
-    xTicks, yTicks = nDim*[None], nDim*[None]
-
+    ##### Define colormap
+    myColorMap = setCustomColorMaps(colors)
 
     ##### Matplotlib and figure settings
     axisColor = '#333333'
@@ -370,7 +384,12 @@ def plotGTC(chains, **kwargs):
     fig = plt.figure(figsize=(figureWidth,figureWidth))
     axV, axH = [], []
 
+    # These are needed to compute the confidence levels
+    nBinsFlat = np.linspace(0., nBins**2, nBins**2)
 
+    # Left and right panel boundaries
+    panelXrange = np.empty((nDim,2))
+    xTicks, yTicks = nDim*[None], nDim*[None]
 
     ########## 2D contour plots
     if not doOnly1dPlot:
@@ -391,7 +410,7 @@ def plotGTC(chains, **kwargs):
                     if truths is not None:
                         truthsForPlot2D = [[truths[k,i], truths[k,j]] for k in range(len(truths))]
                     # Plot!
[…]
