### 2016-08-05 f21a4f5
renamed confidence levels to avoid confusion
