### 2016-08-05 c4df44a
sigma confidence levels
