### 2016-08-16 6699c0e
paper.md : added acknowledgements for funding
