+++ paper/paper.md
@@ -36,4 +36,8 @@ GTC = pygtc.plotGTC(chains=[dat1,dat2,dat3],
 
 `pygtc` relies on `matplotlib` [@matplotlib] and `scipy` [@scipy] (optional) to produce aesthetically appealing GTCs, and is designed to accept the standard output of the `emcee` package or any other likelihood sampler that returns a list of (weighted or unit-weighted) sample points. The source code for `pygtc` is available at Zenodo [@pygtc_zenodo] and development is ongoing at GitHub [@pygtc].
 
+# Acknowledgements
+
+Work at Argonne National Laboratory was supported under the U.S. Department of Energy contract DE-AC02-06CH11357.
+
 # References
