+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup
 import pygtc
 
 setup(
-    name = 'pyGTC',
+    name = 'pygtc',
     description = 'Make an awesome giant triangle confusogram (gtc)!',
     version = '0.1.1',
     author = 'Sebastian Bocquet and Faustin Carter',
@@ -11,7 +11,11 @@ setup(
     url = 'http://github.com/sebastianbocquet/pygtc',
     packages = ['pygtc'],
     long_description = open('README.rst').read(),
-    install_requires = ['numpy', 'matplotlib', 'scipy'],
+    install_requires = [
+        'numpy>=1.5',
+        'matplotlib>=1.5',
+        'scipy>=0.14'
+    ],
     classifiers = [
         'Development Status :: 4 - Beta',
         'Intended Audience :: Science/Research',
