### 2016-09-02 188fe66
docstring edit
