+++ pygtc/pygtc.py
@@ -10,7 +10,7 @@ try:
     haveScipy = True
 except ImportError:
     haveScipy = False
-    
+
 __all__ = ['plotGTC']
 
 
@@ -30,6 +30,8 @@ def plotGTC(chains, **kwargs):
         (http://dan.iel.fm/emcee/current/) - and you should! - each element
         of chains is an ``EnsembleSampler.flatchain`` object.
 
+    Keyword Arguments
+    -----------------
     weights : array-like[nSamples] or a list[[nSamples1], ...]
         Weights for the sample points. If a 1d array is passed, the same
         weights are used for all dimension in chains. If a list of 1d arrays
@@ -71,7 +73,7 @@ def plotGTC(chains, **kwargs):
         User-defined line styles for the truth lines, must be one per set of
         truths passed to `truths`. Default line styles
         are ``['--',':','dashdot']``.
-    
+
     priors : list of tuples [(mu1, sigma1), ...]
         Each tuple describes a Gaussian to be plotted over that parameter's
         histogram. The number of priors must equal the number of dimensions
@@ -87,7 +89,7 @@ def plotGTC(chains, **kwargs):
     GaussianConfLevels : bool
         Whether you want non-standard 2d Gaussian "sigma" confidence levels
         instead of the usual 68%, 95%, 99% confidence levels. Default is ``False``.
-    
+
     nBins : int
         An integer describing the number of bins used to compute the
         histograms. Default is 30.
@@ -97,12 +99,12 @@ def plotGTC(chains, **kwargs):
         bins. Default is 1. Set to 0 for no smoothing.
 
     filledPlots : bool
-        Whether you want the 2d contours and the 1d histograms to be 
+        Whether you want the 2d contours and the 1d histograms to be
         filled. Default is ``True``.
-    
+
     plotDensity : bool
         Whether you want to see the 2d density of points. Default is ``False``.
-    
+
     figureSize : float or string
         A number in inches describing the length = width of the GTC, or a
         string indicating a predefined journal setting and whether the
@@ -115,7 +117,7 @@ def plotGTC(chains, **kwargs):
         Options are ``'loose'`` or ``'tight'``. Determines whether there is
         some space between the subplots of the GTC or not. Default is
         ``'tight'``.
-    
+
     legendMarker : string
         Options are ``'All'``, ``'None'``, ``'Auto'``. ``'All'`` and ``'None'``
         force-show or force-hide all label markers. ``'Auto'`` shows label
@@ -139,7 +141,7 @@ def plotGTC(chains, **kwargs):
     doOnly1dPlot : bool
         Plot only ONE 1d histogram. If this is True, then chains must have
         shape ``(samples,1)``. Default is ``False``.
-    
+
     mathTextFontSet : string
         Set consistent font family in rcParams. Default is ``stixsans``. Set
         to ``'None'`` to use the setting in your matplotlib rc.
@@ -176,7 +178,7 @@ def plotGTC(chains, **kwargs):
     figSizeDict = { 'APJ_column' : 245.26653 / mplPPI,
                     'APJ_page' : 513.11743 / mplPPI,
                     'MNRAS_column' : 240. / mplPPI,
-                    'MNRAS_page' : 504. / mplPPI, 
+                    'MNRAS_page' : 504. / mplPPI,
                     'AandA_column' : 256.0748 / mplPPI,
                     'AandA_page' : 523.5307 / mplPPI}
 
@@ -318,7 +320,7 @@ def plotGTC(chains, **kwargs):
     confLevels = (.3173, .0455, .0027)
     if GaussianConfLevels:
         confLevels = (.6065, .1353, .0111)
-        
+
 
     # Data binning and smoothing
[…]
