### 2016-09-02 2003772
pythonized bin center computation
