### 2016-09-02 c599e6a
minor reordering
