### 2016-09-05 883a318
Documentation updates. Preparing for new version.
