+++ docs/conf.py
@@ -20,7 +20,16 @@ import os
 #This is important because RTD doesn't support any c-based modules.
 import mock
 
-MOCK_MODULES = ['numpy', 'scipy', 'matplotlib', 'matplotlib.pyplot', 'matplotlib.ticker', 'scipy.ndimage', 'scipy.stats']
+MOCK_MODULES = ['numpy',
+                'scipy',
+                'matplotlib',
+                'matplotlib.pyplot',
+                'matplotlib.ticker',
+                'matplotlib.colors',
+                'matplotlib.colors.LinearSegmentedColormap',
+                'scipy.ndimage',
+                'scipy.stats']
+
 for mod_name in MOCK_MODULES:
     sys.modules[mod_name] = mock.Mock()
 
@@ -67,9 +76,9 @@ author = u'Sebastian Bocquet and Faustin W. Carter'
 # built documents.
 #
 # The short X.Y version.
-version = u'0.1'
+version = u'0.2'
 # The full version, including alpha/beta/rc tags.
-release = u'0.1.0'
+release = u'0.2.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
