### 2016-09-05 17f6bf5
Documentation updates. Preparing for new version.
