+++ docs/installation.rst
@@ -1,0 +1,50 @@
+============
+Installation
+============
+
+Required packages
+~~~~~~~~~~~~~~~~~
+
+pygtc is written in Python 2.7 and requires the following packages:
+
+* numpy >= 1.5
+* matplotlib >= 1.5
+* scipy (optional)
+
+
+Downloading and installing
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+The latest stable version of scraps is hosted at `PyPi
+<http://pypi.python.org/pypi/pygtc/>`_.
+
+If you like pip, you can install with::
+
+  pip install pygtc
+
+Or, alternatively, download and extract the tarball from the above link and then
+install from source with::
+
+  cd directory_you_unpacked_scraps_into
+  python setup.py install
+
+
+Development
+~~~~~~~~~~~
+
+Development happens at `github <https://github.com/SebastianBocquet/pygtc>`_. You can
+clone the repo with::
+
+  git clone https://github.com/SebastianBocquet/pygtc
+
+And you can install it in developer mode with pip::
+
+  pip install -e directory_where_you_cloned_pygtc
+
+or from source::
+
+  cd directory_cloned_pygtc_into
+  python setup.py develop
+
+If you like this project and want to contribute, send a message over at the
+github repo and you can be added as a collaborator!
