+++ README.rst
@@ -18,7 +18,7 @@ Here's an example of a GTC::
 
 .. image:: ./docs/_static/demo_files/demo_8_0.png
 
-**But doesn't this already exist in ``corner``, ``distUtils``, etc...?***
+**But doesn't this already exist in corner, distUtils, etc...?**
 
 Although several other packages exists to make such a plot, we were unsatisfied
 with the amount of extra work required to massage the result into something we
