+++ README.rst
@@ -8,7 +8,9 @@ displaying the results of a Monte-Carlo Markov Chain (MCMC) sampling or similar
 analysis. (For a discussion of MCMC analysis, see the excellent ``emcee``
 package.) The recovered parameter constraints are displayed on a grid in which
 the diagonal shows the one-dimensional posteriors and the lower-left triangle
-shows the pairwise projections.
+shows the pairwise projections. You might want to look at a plot like this if
+you are fitting model to data and want to see the parameter covariances along
+with the priors.
 
 Here's an example of a GTC::
 
