### 2016-09-06 d8fc7bf
got mathjax working in sphinx and added some more to the docs
