### 2016-09-09 2a03fa9
Added docstrings to private functions.
