+++ pygtc/pygtc.py
@@ -311,13 +311,14 @@ def plotGTC(chains, **kwargs):
         assert __isstr(plotName), "plotName must be a string type"
 
     # Define which confidence levels to show
-    nConfidenceLevels = kwargs.pop('nConfidenceLevels', 2) #How many of the above confidence levels to show
+    nConfidenceLevels = kwargs.pop('nConfidenceLevels', 2) #How many of the confidence levels to show
     assert nConfidenceLevels in [1,2,3], "nConfidenceLevels must be 1, 2, or 3"
 
     # 2d confidence levels: Gaussian or (68%, 95%, 99%)
     GaussianConfLevels = kwargs.pop('GaussianConfLevels', False)
     confLevels = (.3173, .0455, .0027)
     if GaussianConfLevels:
+        #1d confidence levels
         confLevels = (.6065, .1353, .0111)
 
 
@@ -422,7 +423,10 @@ def plotGTC(chains, **kwargs):
                     if truths is not None:
                         truthsForPlot2D = [[truths[k,i], truths[k,j]] for k in range(len(truths))]
                     # Plot!
-                    ax = __plot2d(ax, nChains, chainsForPlot2D, weights, nBins, nBinsFlat, smoothingKernel, filledPlots, colors, nConfidenceLevels, confLevels, truthsForPlot2D, truthColors, truthLineStyles, plotDensity, myColorMap)
+                    ax = __plot2d(ax, nChains, chainsForPlot2D, weights, nBins,
+                                nBinsFlat, smoothingKernel, filledPlots, colors,
+                                nConfidenceLevels, confLevels, truthsForPlot2D,
+                                truthColors, truthLineStyles, plotDensity, myColorMap)
 
                     ##### Range
                     if paramRanges is not None:
@@ -527,7 +531,9 @@ def plotGTC(chains, **kwargs):
                 if priors[i] and priors[i][1]>0:
                     prior1d = priors[i]
             # Plot!
-            ax = __plot1d(ax, nChains, chainsForPlot1D, weights, nBins, smoothingKernel, filledPlots, colors, truthsForPlot1D, truthColors, truthLineStyles, prior1d, priorColor)
+            ax = __plot1d(ax, nChains, chainsForPlot1D, weights, nBins,
+                        smoothingKernel, filledPlots, colors, truthsForPlot1D,
+                        truthColors, truthLineStyles, prior1d, priorColor)
 
 
             ##### Panel layout
@@ -687,7 +693,58 @@ def plotGTC(chains, **kwargs):
 
 #################### Create single 1d panel
 
-def __plot1d(ax, nChains, chains1d, weights, nBins, smoothingKernel, filledPlots, colors, truths1d, truthColors, truthLineStyles, prior1d, priorColor):
+def __plot1d(ax, nChains, chains1d, weights, nBins, smoothingKernel,
+            filledPlots, colors, truths1d, truthColors, truthLineStyles,
+            prior1d, priorColor):
+    r"""Plot the 1d histogram and optional prior.
+
+    Parameters
+    ----------
+    ax : matplotlib.pyplot.axis
+        Axis on which to plot the histogram(s)
+
+    nChains : int
+        How many chains are you passing?
+
+    chains1d : list-like
+        A list of `nChains` 1d chains: [chain1, chain2, etc...]
+
+    weights : list-like
+        A list of `nChains` weights.
+
+    nBins : int
+        How many histogram bins?
+
+    smoothingKernel : int
+        Number of bins to smooth over
+
+    filledPlots : bool
+        Want the area under the curve filled in?
+
+    colors : list-like
+        List of `nChains` tuples. Each tuple must have at least two colors.
+
+    truths1d : list-like
+        List of truths to overplot on the histogram.
+
+    truthColors : list-like
[…]
