### 2016-09-12 00dc034
Option to disable tick label rotation
