### 2016-09-12 2fc207d
Option to disable tick label rotation - fixed
