### 2016-09-12 3e88377
make alignment of rotated tick labels depend on label length
