### 2016-09-13 99e9d15
some fixes in the docs
