### 2016-09-13 cb0ab1e
Version number is now in VERSION.txt and only there. Everything else pulls from there.
