+++ docs/conf.py
@@ -76,10 +76,16 @@ author = u'Sebastian Bocquet and Faustin W. Carter'
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
+
+version_file = open(os.path.join('..', 'VERSION.txt'))
+version_number = version_file.read().strip()
+version_file.close()
+
 # The short X.Y version.
-version = u'0.2'
+v_main, v_sub, v_release = version_number.split('.')
+version = v_main + '.' + v_sub
 # The full version, including alpha/beta/rc tags.
-release = u'0.2.0'
+release = version
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
