### 2016-09-13 17edc06
added VERSION.txt
