### 2016-09-13 18ce8a4
Added custom font control and also sorted out the deal with rendering LaTex with custom font-faces
