+++ pygtc/pygtc.py
@@ -87,8 +87,8 @@ def plotGTC(chains, **kwargs):
         or 3. Default is 2.
 
     GaussianConfLevels : bool
-        Whether you want 2d Gaussian "sigma" confidence levels instead of the
-        standard 68%, 95%, 99% confidence levels. Default is ``False``.
+        Whether you want 2d Gaussian "sigma" confidence levels (39%, 86%, 99%) instead of the
+        standard 1d confidence levels (68%, 95%, 99%). Default is ``False``.
 
     nBins : int
         An integer describing the number of bins used to compute the
@@ -127,7 +127,7 @@ def plotGTC(chains, **kwargs):
         Set the boundaries of each paramter range. Must provide a tuples for
         each dimension of `chains`. If ``None`` is provided for a
         parameter, the range defaults to the width of the histogram.
-    
+
     labelRotation : tuple [2]
         Rotate the tick labels by 45 degrees for less overlap. Can be set for
         the x- and y-axis separately. Options are ``(True,True)``,
@@ -149,8 +149,24 @@ def plotGTC(chains, **kwargs):
         shape ``(samples,1)``. Default is ``False``.
 
     mathTextFontSet : string
-        Set consistent font family in rcParams. Default is ``stixsans``. Set
-        to ``'None'`` to use the setting in your matplotlib rc.
+        Set font family for rendering LaTex. Default is ``'stixsans'``. Set
+        to ``None`` to use the default setting in your matplotlib rc. See Notes
+        for known issues regarding this keyword.
+
+    customLabelFont : ``matplotlib.fontdict``
+        Full customization of label fonts. See matplotlib for full
+        documentation. Default: ``customLabelFont = {'size':9}``
+
+    customLegendFont : ``matplotlib.fontdict``
+        Full customization of legend fonts. See matplotlib for full
+        documentation. Default: ``customLabelFont = {'size':9}``
+
+    holdRC : bool
+        Whether or not to reset rcParams back to default. You may wish to set
+        this to ``False`` if you are working in interactive mode (ie with
+        IPython or in a JuPyter notebook) and you want the plots that display to
+        be identical to the plots that save in the pdf. See Notes below for more
+        information.
 
     Returns
     -------
@@ -159,6 +175,28 @@ def plotGTC(chains, **kwargs):
         customization after it gets returned. If you are using a ``JuPyter``
         notebook with inline plotting enabled, you should assign a variable
         to catch the return or else the figure will plot twice.
+
+    Note
+    ----
+    If you are calling ``plotGTC`` from within an interactive python session (ie
+    via IPython or in a JuPyter notebook), the label font in the saved pdf may
+    differ from the plot that appears when calling ``matplotlib.pyplot.show()``.
+
+    This will happen if the mathTextFontSet keyword sets a value that is
+    different than the one stored in ``rcParams['mathtext.fontset']`` and you
+    are using equations in your labels by enclosing them in $..$. The output pdf
+    will display correctly, but the interactive plot will use whatever is stored
+    in the rcParams default to render the text that is inside the $..$.
+    Unfortunately, this is an oversight in matplotlib's design, which only
+    allows one global location for specifying this setting. As a workaround, you
+    can set ``holdRC = True`` when calling ``plotGTC`` and it will *not* reset
+    your rcParams back to their default state. Thus, when the figure renders in
+    interactive mode, it will match the saved pdf. If you wish to reset your
+    rcParams back to default at any point, you can call
[…]
