### 2016-09-13 0b4c818
fix customLegendFont for legend
