### 2016-09-13 8487857
Small fixes in docstring
