+++ README.rst
@@ -9,7 +9,7 @@ analysis. (For a discussion of MCMC analysis, see the excellent ``emcee``
 package.) The recovered parameter constraints are displayed on a grid in which
 the diagonal shows the one-dimensional posteriors (and, optionally, priors) and
 the lower-left triangle shows the pairwise projections. You might want to look
-at a plot like this if you are fitting model to data and want to see the
+at a plot like this if you are fitting a model to data and want to see the
 parameter covariances along with the priors.
 
 Here's an example of a GTC with some random data and arbitrary labels::
