### 2016-09-13 17d72b8
Updates to paper refs
