+++ paper/paper.bib
@@ -86,3 +86,16 @@ archivePrefix = "arXiv",
   note = {[Online; accessed 2016-07-27]},
   doi = {10.1109/MCSE.2011.37}
 }
+
+@article{numpy,
+   author = "Walt, St{\'e}fan van der and Colbert, S. Chris and Varoquaux, Ga{\"e}l",
+   title = "The NumPy Array: A Structure for Efficient Numerical Computation",
+   journal = "Computing in Science & Engineering",
+   year = "2011",
+   volume = "13",
+   number = "2",
+   eid = ,
+   pages = "22-30",
+   url = "http://scitation.aip.org/content/aip/journal/cise/13/2/10.1109/MCSE.2011.37",
+   doi = "http://dx.doi.org/10.1109/MCSE.2011.37" 
+}
