### 2016-09-14 cf42888
Monkey patched the .set_x method of tick labels to allow for shifting them around. Then shifted them in a nice way. Also added an option to shift them a little more or less if desired. Also fixed default fonts and updated documentation for all changes.
