+++ pygtc/pygtc.py
@@ -1,9 +1,10 @@
 from matplotlib import pyplot as plt
 from matplotlib.colors import LinearSegmentedColormap
-#from matplotlib.colors import LogNorm
+import matplotlib as mpl
 import matplotlib.colors as mplcolors
 import numpy as np
 import matplotlib.ticker as mtik
+import types
 try:
     import scipy.ndimage
     from scipy.stats import norm
@@ -134,6 +135,14 @@ def plotGTC(chains, **kwargs):
         ``(True,False)``, ``(False,True)``, ``(False,False)``, ``None``.
         Using ``None`` sets to default ``(True,True)``.
 
+    tickShifts : tuple [2]
+        Shift the x/y tick labels horizontally/vertically by a fraction of the
+        tick spacing. Example tickShifts = (0.1, 0.05) shifts the x-tick labels
+        right by ten percent of the tick spacing and shifts the y-tick labels up
+        by five percent of the tick spacing. Default is (0.1, 0.1). If tick
+        rotation is turned off for either axis, then the corresponding shift is
+        set to zero.
+
     colorsOrder : list-like[nDims]
         The color order for chains passed to `chains`. Default is
         ``['blues', 'greens', 'yellows', 'reds', 'purples']``. Currently,
@@ -155,11 +164,15 @@ def plotGTC(chains, **kwargs):
 
     customLabelFont : ``matplotlib.fontdict``
         Full customization of label fonts. See matplotlib for full
-        documentation. Default: ``customLabelFont = {'size':9}``
+        documentation. Default is ``{'family':'Arial', 'size':9}``.
 
     customLegendFont : ``matplotlib.fontdict``
         Full customization of legend fonts. See matplotlib for full
-        documentation. Default: ``customLabelFont = {'size':9}``
+        documentation. Default is ``{'family':'Arial', 'size':9}``.
+
+    customTickFont : ``matplotlib.fontdict``
+        Full customization of tick label fonts. See matplotlib for full
+        documentation. Default is ``{'family':'Arial', 'size':6}``.
 
     holdRC : bool
         Whether or not to reset rcParams back to default. You may wish to set
@@ -298,6 +311,15 @@ def plotGTC(chains, **kwargs):
     # Rotated tick labels
     labelRotation = kwargs.pop('labelRotation', (True,True))
 
+    # Shifted tick labels, Default is nudge by 0.1 * tick spacing
+    shiftX, shiftY = kwargs.pop('tickShifts', (0.1, 0.1))
+
+    #If the rotation is turned off, then don't shift the labels
+    if not labelRotation[0]:
+        shiftX = 0
+    if not labelRotation[1]:
+        shiftY = 0
+
     # User-defined color ordering
     customColorsOrder = kwargs.pop('colorsOrder', None) #Labels for multiple chains, goes in plot legend
     if customColorsOrder is not None:
@@ -430,19 +452,36 @@ def plotGTC(chains, **kwargs):
     oldMathTextFontSet = plt.rcParams['mathtext.fontset']
     if mathTextFontSet is not None:
         plt.rcParams['mathtext.fontset'] = mathTextFontSet
-        
+
     holdRC = kwargs.pop('holdRC', False)
     assert holdRC in [True, False], "holdRC must be True or False."
 
     #Grab the custom fontdicts
     #Default size is 9 for all labels.
-    defaultFontSize = 9
+    defaultFontFamily = 'Arial'
+    defaultLabelFontSize = 9
+    defaultTickFontSize = 6
+
     customLabelFont = kwargs.pop('customLabelFont', {})
     if 'size' not in customLabelFont.keys():
-        customLabelFont['size'] = defaultFontSize
[…]
