### 2016-09-14 96586b2
Pinned matplotlib version to less than 2.0 as that release will definitely break pygtc in many ways.
