+++ setup.py
@@ -18,7 +18,7 @@ setup(
     long_description = open('README.rst').read(),
     install_requires = [
         'numpy>=1.5',
-        'matplotlib>=1.5'
+        'matplotlib>=1.5, <2.0'
     ],
     classifiers = [
         'Development Status :: 4 - Beta',
