+++ docs/installation.rst
@@ -8,7 +8,7 @@ Required packages
 pygtc is written in Python 2.7 and requires the following packages:
 
 * numpy >= 1.5
-* matplotlib >= 1.5
+* matplotlib >= 1.5 and < 2.0
 * scipy (optional)
 
 
