### 2016-09-14 6c62e8d
added note about CL vs sigma contour levels
