+++ docs/index.rst
@@ -27,6 +27,19 @@ labels):
 
 .. image:: _static/demo_files/demo_8_0.png
 
+**Note about contour levels**
+In the above example, you see two sets of contour levels. These show the 68%
+and 95% *confidence levels* (in addition, pygtc could also show the 99% confidence
+level if you want). Note that these are indeed the 68% and 95% confidence levels,
+and not Gaussian 1, 2, and 3 sigma levels. Your posterior distributions will in
+general not be Gaussian, and so we argue that there is no "right" choice. We
+therefore recommend that you stick to the definition that is
+commonly used in your field, and clearly mention what quantities you are showing
+in your figure caption. You can switch to displaying Gaussian sigma levels
+(in two dimensions, these correspond to the 39%, 86%, and 99% confidence levels)
+by setting `GaussianConfLevels` to `True`.
+
+
 Contents:
 ---------
 .. toctree::
