### 2016-09-14 6bc8ef3
Slight rewording of note about CL
