### 2016-09-15 ffbaa86
Added Planck/WMAP example to docs
