### 2016-09-15 30e968f
Small change in readme
