+++ README.rst
@@ -3,7 +3,7 @@ pygtc.py
 
 **What is a Giant Triangle Confusogram?**
 
-A Giant-Triangle-Confusogram (GTC, aka triangle/corner plot) is a way of
+A Giant-Triangle-Confusogram (GTC, aka triangle plot) is a way of
 displaying the results of a Monte-Carlo Markov Chain (MCMC) sampling or similar
 analysis. (For a discussion of MCMC analysis, see the excellent ``emcee``
 package.) The recovered parameter constraints are displayed on a grid in which
@@ -20,7 +20,7 @@ Here's an example of a GTC with some random data and arbitrary labels::
 
 .. image:: ./docs/_static/demo_files/demo_8_0.png
 
-**But doesn't this already exist in corner, distUtils, etc...?**
+**But doesn't this already exist in corner.py, distUtils, etc...?**
 
 Although several other packages exists to make such a plot, we were unsatisfied
 with the amount of extra work required to massage the result into something we
