### 2016-09-15 cc6bd63
Small change in docs index
