### 2016-09-15 f92bad2
Fixed the example plots to have correct resolution
