+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 Version 0.2.0:
   * Option to turn off tick label rotation
   * Long tick labels are now lined up correctly with their ticks
+  * Option to fine tune position of label relative to tick
   * Can choose to use 1d or 2d Gaussian sigma contours
   * scipy is now an optional dependency
   * Can choose to fill contours and 1d histograms independently
@@ -8,6 +9,7 @@ Version 0.2.0:
   * Full control over fonts for labels, ticks, and legend
   * Use built-in matplotlib LaTex renderer for equations in labels so no LaTex
     installation is required locally.
+  * Added a new example to the documentation
 
 Version 0.1.1:
   * Prettier truth colors
