### 2016-09-15 7eb2a27
added DOI to docs
