+++ docs/index.rst
@@ -60,3 +60,6 @@ Please report any problems or pitch new ideas on `GitHub
 <https://github.com/SebastianBocquet/pygtc/issues>`__ where pygtc is being
 developed. Bugs will be squashed ASAP and feature requests will be seriously
 considered.
+
+.. image:: https://zenodo.org/badge/DOI/10.5281/zenodo.154122.svg
+   :target: https://doi.org/10.5281/zenodo.154122
