### 2016-09-16 #46 @SebastianBocquet
Dear All,

below you will find our response to the referee report, highlighted in
green.

Best regards,

Sebastian Bocquet

----

Authors' response:

We would like to thank Dan for carefully reviewing our software package and
paper. He points out several important issues. We have addressed each of
the raised concerns, and made changes as described below.

General checks

Version: Does the release version given match the GitHub release (0.1.1)?

There is no archived release linked from the documentation or readme and
there is no DOI given in the paper.

We have version bumped to 0.2.0 to generate a Zenodo DOI, which is:

10.5281/zenodo.154122

We added the link to the readme, documentation, and paper.

Functionality

Installation: Does installation proceed as outlined in the documentation?

The installation instructions are missing a list of prerequisites. Does
scipy really need to be a hard dependency?

We have made scipy optional, updated the installation instructions, and
defined minimum (and maximum) versions (to avoid forced upgrades of numpy
and/or matplotlib).

Documentation

A statement of need: Do the authors clearly state what problems the
software is designed to solve and who the target audience is?

There is a vague statement of need in the README but it's not very clear
and it is completely missing from the docs. The contents of the paper (and
maybe more discussion) should be included in the docs.

We have updated both the readme.rst in the repo, as well as the
documentation to address this.

There is also no discussion of how to interpret these plots. For example,
the (incorrect, I would argue
<http://corner.readthedocs.io/en/latest/pages/sigmas.html>!) choice of
contour levels should be stated and justified.

We added a note about confidence levels for 2d contours to the
documentation. As posterior distributions are not necessarily Gaussian, we
argue that there is no right or wrong in choosing Gaussian “sigma” or 68%
and 95% confidence levels, as long as it is clear which definition is used.
However, we did add the option to show the “sigma” levels.

Installation instructions: Is there a clearly-stated list of dependencies?
Ideally these should be handled with an automated package management
solution.

In principle, the pip command could install all of the dependencies but
they should be discussed in the installation instructions.

They are now clearly stated, along with min/max versions to avoid pip
forcing an upgrade of packages that don’t need to be.

Example usage: Do the authors include examples of how to use the software
(ideally to solve real-world analysis problems).

There is a demo notebook included in the repository but the docs have no
discussion of the usage or even a link to the demo. I think that a gallery
of examples or at least a tutorial is required.

The notebook has been extended to the documentation, where it is the basis
of a tutorial for how to use pygtc. We also added a second example showing
off pygtc with some publicly available cosmological data.

Functionality documentation: Is the core functionality of the software
documented to a satisfactory level (e.g. API method documentation)?

The entry point is documented by its docstring and an auto-generated docs
entry. This isn't really sufficient to constitute documentation.

We have added two tutorials to the documentation, which should be
sufficient to get a new user up and running quickly and demo all of the
major features. For a detailed description of the options, the docstring is
the logical place to turn, and we have carefully documented every option in
the primary user-facing function, and we give a description of each
argument in the helper functions (which are not meant to be user-facing).

Automated tests: Are there automated tests or manual steps described so
that the function of the software can be verified?

No tests are currently included. It might be worth checking out the tests
from corner.py <https://github.com/dfm/corner.py> for an example of how
this could be implemented for similar software.

We understand that from the development perspective an automated test could
be useful to make sure something isn’t broken when a bug is fixed. However,
for a user looking to verify the utility of the package, a test that
compares two images to see if they are the same at the pixel level may not
be terribly useful. This is because anyone may have a different rcParams
default, there is no way we are going to exhaustively override every
possible rcParam, and it only takes one difference to generate an image
that isn’t pixel-for-pixel the same as the one we made on our system (i.e.,
they set their default font color to dark grey, and we don’t specify that
default). This scenario would generate a “failed” test, even though the
user is getting what they want.

A better indication of whether the package works is for the user to run the
included ipython notebook tutorial and simply look at the plots it
generates.

Community guidelines: Are there clear guidelines for third parties wishing
to 1) Contribute to the software 2) Report issues or problems with the
software 3) Seek support

Missing. There should (at least!) be a link to the GitHub repository from
the docs.

This has been fixed.

Software paper

A statement of need: Do the authors clearly state what problems the
software is designed to solve and who the target audience is?

This is well described but the paper incorrectly states that corner.py
can't overplot multiple samples. This has never been true!

We have removed the erroneous statement about corner.py.

The claim is also made that the resulting plots are compatible with
different journals and a list of options are given but it's not clear to me
why this is necessary and why the dpi enters - no one should be submitting
rasterized figures to the journals!

We agree that rasterized figures are basically the devil, which is why
pygtc saves to pdf! However, when one chooses a particular font size, that
font size is rendered relative to the size in inches of the figure, so now
we have to choose a size in inches so the font knows how big to be. At some
point a reference is required to get the relative size of the axes and the
fonts right, and so we use the dpi as a measuring stick. We believe that
the figure size should be set such that the font size is what you expect
(and what the journal wants) and that the figure may be inserted into the
journal without needing to tweak width/height settings in LaTex. Because of
this, we give you precisely what you want when you specify a custom font
size.

References: Do all archival references that should have a DOI list one
(e.g. papers, datasets, software)?

The DOI for numpy/scipy is 10.1109/MCSE.2011.37 and matplotlib is
10.1109/MCSE.2007.55. ref

These are cited correctly now.

Other comments

Packages should never alter the matplotlib rc. The fact that plots made
after calling pygtc are different from the same plots made before is not ok.

In hindsight, we couldn’t agree more. We have completely updated pygtc to
function without touching matplotlib rc (except for one case, see below).

The one case: When using matplotlib’s built in Tex parser to convert LaTex
expressions into math, the only place to specify the font family used for
the math is in the ‘mathtext.fontset’ rcParam. This is in contrast to any
other text in matplotlib, which keeps its own fontdict separate from the
rcParams. (We have opened issue #7107 over at matplotlib’s github repo to
notify them of this and they have added the fix to their 2.0 release
milestones. We will update pygtc again when matplotlib 2.0 comes out.)

To get around this, we change the ‘mathtext.fontset’ rcParam, save the
plot, revert the rcParam back to whatever it was, and then return the
figure to the caller. Unfortunately, in interactive mode (i.e., a notebook
or IPython), the figure will now render to the screen using whatever is in
the rcParams, and may be different from what was desired. We thus include a
flag that will turn off resetting back to original rcParams. If the user
specifies holdRC=True, then the ‘mathtext.fontset’ rcParam will not be
reset back to the default, allowing the caller to render the figure
properly.

On Mon, Aug 1, 2016 at 1:50 PM, Dan Foreman-Mackey <notifications@github.com
> wrote:

> General checks
>
> Version: Does the release version given match the GitHub release (0.1.1)?
>
> There is no archived release linked from the documentation or readme and
> there is no DOI given in the paper.
> Functionality
>
> Installation: Does installation proceed as outlined in the documentation?
>
> The installation instructions are missing a list of prerequisites. Does
> scipy really need to be a hard dependency?
> Documentation
>
> A statement of need: Do the authors clearly state what problems the
> software is designed to solve and who the target audience is?
>
> There is a vague statement of need in the README but it's not very clear
> and it is completely missing from the docs. The contents of the paper (and
> maybe more discussion) should be included in the docs. There is also no
> discussion of how to interpret these plots. For example, the (incorrect, I
> would argue <http://corner.readthedocs.io/en/latest/pages/sigmas.html>!)
> choice of contour levels should be stated and justified.
>
> Installation instructions: Is there a clearly-stated list of dependencies?
> Ideally these should be handled with an automated package management
> solution.
>
> In principle, the pip command could install all of the dependencies but
> they should be discussed in the installation instructions.
>
> Example usage: Do the authors include examples of how to use the software
> (ideally to solve real-world analysis problems).
>
> There is a demo notebook included in the repository but the docs have no
> discussion of the usage or even a link to the demo. I think that a gallery
> of examples or at least a tutorial is required.
>
> Functionality documentation: Is the core functionality of the software
> documented to a satisfactory level (e.g. API method documentation)?
>
> The entry point is documented by its docstring and an auto-generated docs
> entry. This isn't really sufficient to constitute documentation.
>
> Automated tests: Are there automated tests or manual steps described so
> that the function of the software can be verified?
>
> No tests are currently included. It might be worth checking out the tests
> from corner.py <https://github.com/dfm/corner.py> for an example of how
> this could be implemented for similar software.
>
> Community guidelines: Are there clear guidelines for third parties wishing
> to 1) Contribute to the software 2) Report issues or problems with the
> software 3) Seek support
>
> Missing. There should (at least!) be a link to the GitHub repository from
> the docs.
> Software paper
>
> A statement of need: Do the authors clearly state what problems the
> software is designed to solve and who the target audience is?
>
> This is well described but the paper incorrectly states that corner.py
> can't overplot multiple samples. This has never been true! 😉 The claim
> is also made that the resulting plots are compatible with different
> journals and a list of options are given but it's not clear to me why this
> is necessary and why the dpi enters - no one should be submitting
> rasterized figures to the journals!
>
> References: Do all archival references that should have a DOI list one
> (e.g. papers, datasets, software)?
>
> The DOI for numpy/scipy is 10.1109/MCSE.2011.37 and matplotlib is
> 10.1109/MCSE.2007.55. ref <https://www.scipy.org/citing.html>
> Other comments
>
> Packages should *never* alter the matplotlib rc. The fact that plots made
> *after* calling pygtc are different from the same plots made before is
> not ok.
>
