### 2016-09-19 #46 @tacaswell
It is probably too late, but mpl has  `mpl.rc_context` and `mpl.style.context`  which both provide context managers for setting and restoring rcParams.

> This is because anyone may have a different rcParams
> default, there is no way we are going to exhaustively override every
> possible rcParam,

This is actually pretty easy to do

```python
mpl.rcdefaults()
mpl.style.use(...)
```

https://github.com/matplotlib/pytest-mpl and the image comparison tests in mpl both include a way to specify rcparams as part of the test and ensure they are properly cleaned up.
