### 2016-09-19 #46 @tacaswell
https://github.com/matplotlib/matplotlib/issues/7107 xref to mpl issue
