### 2016-09-19 #46 @arfon
> @arfon: is it possible to update the software version number and rebuild the paper?

@dfm - sure thing. Here's the new paper: [10.21105.joss.00046.pdf](https://github.com/openjournals/joss-reviews/files/481100/10.21105.joss.00046.pdf)
