### 2016-09-24 4f8106b
Starting to write some tests
