+++ tests/test_plotGTC.py
@@ -1,0 +1,8 @@
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+from matplotlib.testing.decorators import image_comparison
+
+import pygtc
+
+#Tests?
