### 2016-09-28 a153ef7
renamed GaussianConfLevels to gaussianConfLevels
