### 2016-09-28 6de11c8
Added code to deprecate GaussianConfLevels
