### 2016-09-28 0bdf0ad
Added unit tests for almost every kwarg
