+++ tests/test_plotGTC.py
@@ -1,8 +1,0 @@
-import numpy as np
-import pandas as pd
-import matplotlib.pyplot as plt
-from matplotlib.testing.decorators import image_comparison
-
-import pygtc
-
-#Tests?
