+++ pygtc/tests/test_plotGTC.py
@@ -1,0 +1,200 @@
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.testing.decorators import image_comparison
+
+try:
+    import pandas as pd
+    HAS_PANDAS = True
+except ImportError:
+    HAS_PANDAS = False
+
+import pygtc
+
+
+#Set up some global variables for testing
+def _make_random_chain(ndim=4, Npoints=10000, seed = 0):
+    np.random.seed(seed)
+    means = np.random.rand(ndim)
+    cov = .5 - np.random.rand(ndim**2).reshape((ndim,ndim))
+    cov = np.triu(cov)
+    cov += cov.T - np.diag(cov.diagonal())
+    cov = np.dot(cov,cov)
+    samples =  np.random.multivariate_normal(means, cov, Npoints)
+    return samples
+
+# Create two sets of sample points with 4 parameters and 10000 points
+SAMPLES_1 = 2*_make_random_chain(seed = 1)
+SAMPLES_2 = 1+_make_random_chain(seed = 2)
+SAMPLES_1[:,3]+=1e8
+SAMPLES_2[:,3]+=1e8
+
+
+#If this one fails, something is really wrong with matplotlib
+@image_comparison(baseline_images=['img'], extensions=['png'])
+def test_img():
+    fig = plt.figure()
+    ax = fig.add_subplot(111)
+    ax.plot(1,1)
+
+#A test for (almost) every keyword argument
+@image_comparison(baseline_images=['bare'], extensions=['png'])
+def test_GTC_bare():
+    pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2])
+
+@image_comparison(baseline_images=['pandas'], extensions=['png'])
+def test_GTC_pandas():
+    namesNoTex = ['param name', 'B_labmda', 'C', 'lambda']
+
+    if HAS_PANDAS:
+        samples1_pd = pd.DataFrame(SAMPLES_1, columns=namesNoTex)
+        samples2_pd = pd.DataFrame(SAMPLES_2, columns=namesNoTex)
+    else:
+        samples1_pd = SAMPLES_1
+        samples2_pd = SAMPLES_2
+
+    pygtc.plotGTC(chains=[samples1_pd,samples2_pd])
+
+@image_comparison(baseline_images=['paramNames_noTex'], extensions=['png'])
+def test_GTC_paramNames_noTex():
+    namesNoTex = ['param name', 'B_labmda', 'C', 'lambda']
+    pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2],
+                            paramNames = namesNoTex)
+
+@image_comparison(baseline_images=['paramNames_withTex'], extensions=['png'])
+def test_GTC_paramNames_withTex():
+    namesWithTex = ['param name', '$B_\mathrm{\lambda}$', '$Q^a$', '$\\lambda$']
+    pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2],
+                            paramNames = namesWithTex)
+
+@image_comparison(baseline_images=['chainLabels_noTex'], extensions=['png'])
+def test_GTC_chainLabels_noTex():
+    chainLabelsNoTex = ['data1', 'data 2']
+    pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2],
+                            chainLabels = chainLabelsNoTex)
+
+@image_comparison(baseline_images=['chainLabels_withTex'], extensions=['png'])
+def test_GTC_chainLabels_withTex():
+    chainLabelsWithTex = ['data1 $\lambda$', 'data 2']
+    pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2],
[…]
