### 2016-09-28 7145ae9
added testing language to docs
