+++ docs/installation.rst
@@ -46,5 +46,28 @@ or from source::
   cd directory_cloned_pygtc_into
   python setup.py develop
 
+Running tests
+~~~~~~~~~~~~~
+
+For tests to *for sure* run properly, you'll want to have matplotlib v1.5.3
+installed, as they fixed a bug in their ``image_comparison`` decorator. In any
+case, you'll need ``nose`` installed to run the tests. There are two ways to do
+the test suite. You can use the nosetests utility::
+
+  nosetests directory_cloned_pygtc_into
+
+Or, you can run the tests as a script::
+
+  cd directory_cloned_pygtc_into/pygtc/tests
+  python test_plotGTC.py
+
+In either case, there are 25 tests to run, and it should take between 20-30
+seconds to run them all. If the first test fails, there is likely something
+wrong with matplotlib in general.
+
+Contribution and/or bug reports
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
 If you like this project and want to contribute, send a message over at the
-github repo and get involved!
+gitHub repo and get involved. If you find a bug, please open an issue at gitHub.
+Better yet, you can try and fix the bug and submit a pull request!
