### 2016-09-30 f363a73
gaussianConfLevels now called sigmaContourLevels; changed docs accordingly
