### 2016-09-30 243e0ce
replacing 'confidence' with 'contour'
