+++ pygtc/pygtc.py
@@ -83,7 +83,7 @@ def plotGTC(chains, **kwargs):
     plotName : string
         A path to save the GTC to in pdf form. Default is ``None``.
 
-    nConfidenceLevels : int
+    nContourLevels : int
         The number of contour levels to plot in the 2d histograms. May be 1, 2,
         or 3. Default is 2.
 
@@ -384,18 +384,25 @@ def plotGTC(chains, **kwargs):
     if plotName is not None:
         assert __isstr(plotName), "plotName must be a string type"
 
-    # Define which confidence levels to show
-    nConfidenceLevels = kwargs.pop('nConfidenceLevels', 2) #How many of the confidence levels to show
-    assert nConfidenceLevels in [1,2,3], "nConfidenceLevels must be 1, 2, or 3"
+    # Which contour levels to show
+    nContourLevels = kwargs.pop('nContourLevels', 2)
+    assert nContourLevels in [1,2,3], "nContourLevels must be 1, 2, or 3"
+    
+    # Maintain support for older naming convention. TODO: Remove in next major version
+    deprecated_nContourLevels = kwargs.pop('nConfidenceLevels', False)
+    if deprecated_ConfLevels:
+        import warnings
+        warnings.warn("nConfidenceLevels has been replaced by nContourLevels", DeprecationWarning)
+        nContourLevels = deprecated_ConfLevels
+        assert nContourLevels in [1,2,3], "nContourLevels must be 1, 2, or 3"
 
-    # 2d confidence levels: sigma or (68%, 95%, 99%)
+    # 2d contour levels: (68%, 95%, 99%) or sigma (39%, 86%, 99%)
     confLevels = (.3173, .0455, .0027)
     sigmaContourLevels = kwargs.pop('sigmaContourLevels', False)
     if sigmaContourLevels:
-        #1d confidence levels
         confLevels = (.6065, .1353, .0111)
 
-    #Maintain support for older naming convention. TODO: Remove in next major version
+    # Maintain support for older naming convention. TODO: Remove in next major version
     deprecated_ConfLevels = kwargs.pop('gaussianConfLevels', False)
     if deprecated_ConfLevels:
         import warnings
@@ -539,7 +546,7 @@ def plotGTC(chains, **kwargs):
                         truthsForPlot2D = [[truths[k,i], truths[k,j]] for k in range(len(truths))]
                     # Plot!
                     ax = __plot2d(ax, nChains, chainsForPlot2D, weights, nBins,
-                                smoothingKernel, filledPlots, colors, nConfidenceLevels,
+                                smoothingKernel, filledPlots, colors, nContourLevels,
                                 confLevels, truthsForPlot2D, truthColors, truthLineStyles,
                                 plotDensity, myColorMap)
 
@@ -984,7 +991,7 @@ def __plot1d(ax, nChains, chains1d, weights, nBins, smoothingKernel,
 #################### Create single 2d panel
 
 def __plot2d(ax, nChains, chains2d, weights, nBins, smoothingKernel,
-            filledPlots, colors, nConfidenceLevels, confLevels, truths2d,
+            filledPlots, colors, nContourLevels, confLevels, truths2d,
             truthColors, truthLineStyles, plotDensity, myColorMap):
     r"""Plot a 2D histogram in a an axis object and return the axis with plot.
 
@@ -1014,14 +1021,14 @@ def __plot2d(ax, nChains, chains2d, weights, nBins, smoothingKernel,
         Just contours, or filled contours?
 
     colors : list-like
-        List of `nChains` tuples. Each tuple must have at least nConfidenceLevels
+        List of `nChains` tuples. Each tuple must have at least nContourLevels
         colors.
 
[…]
