### 2016-10-01 411ba52
bugfix in deprecation warning
