### 2016-10-01 d4185f9
replace deprecated options in test script
