### 2016-10-01 19b7d56
add requirements file and update docs
