+++ MANIFEST.in
@@ -1,3 +1,4 @@
 include README.rst
 include LICENSE.txt
 include VERSION.txt
+include requirements.txt
