### 2016-10-02 3754019
new DOI for v0.2.1
