+++ paper/paper.bib
@@ -47,8 +47,8 @@
   author = {Sebastian Bocquet and Faustin W. Carter},
   Title = {pygtc Zenodo archive},
   year = 2016,
-  url = {http://doi.org/10.5281/zenodo.154122},
-  doi = {10.5281/zenodo.154122}
+  url = {http://doi.org/10.5281/zenodo.159091},
+  doi = {10.5281/zenodo.159091}
 }
 
 @online{getdist,
