### 2016-10-03 #46 @SebastianBocquet
Dear All,

below you will find our response to the second referee report.

Best regards,
Sebastian Bocquet

_____

Thanks very much for the additional comments. We think our package is much
better as a result of this review and we are both glad we went through this
process. We copy the most recent comments here, and reply to them in green
text.

The comment about "sigmas" has nothing to do with Gaussians. It has to do
with the relationship between the contours that you draw in 2D and the
quantiles that people tend to label in 1D. If you use 68% as a quantile in
1D then you shouldn't use the same in 2D because the contours will look
broader than they should!! I agree that there's no reason to choose a
specific number and encouraging people to say what the used is the best
practice but I'd be careful about the default value that you choose. I'd be
willing to make a large bet that when the vast majority of readers look at
plots like this they interpret the levels as 1 and 2 "sigma" even if you
say "68% contours" and they would be wrong. Also, "confidence" interval
won't, in general, be the correct term – if this is meant to display the
results of MCMC then these are actually credible regions
<https://en.wikipedia.org/wiki/Credible_interval>, for example. With both
of these points in mind, I'd update the terminology ("confidence level" ->
"contour level", "Gaussian" -> None) in the discussion and in the
docstring. (Sorry to harp on this so much but I've spent a lot of time
thinking about this and correcting people's misconceptions about these
types of plots over the years!)

We agree that the discussion has nothing to do with Gaussians, and thank
you for pointing us to the correct use of the confidence/credible terms. We
have refactored the code and changed the variables (and keyword args)
gaussianConfLevels to sigmaContourLevels, and nConfidenceLevels to
nContourLevels. We also have also updated our note in the documentation to
be more precise regarding this. However, we are keeping our defaults set to
68%, 95%, 99%. This is because (at least in our field) we mostly see
contours plotted this way, as this is what getdist produces by default, and
we want to make switching over to pygtc as easy as possible.

It's not actually true that adding version numbers to the setup.py will not
force upgrades (and it's actually generally considered bad practice to put
version numbers in the setup.py
<https://caremad.io/posts/2013/07/setup-vs-requirement/>). If tests were
included then it would be obvious which versions of the packages were
actually required 😉

We have removed the version numbers from setup.py and placed them into a
requirements.txt file instead.

I totally disagree on the point about tests. The point of tests for
something like this is to confirm that the results of the sample plots are
the same for many different versions of numpy and matplotlib (trust me,
this is not a given – adding tests to corner.py found a few bugs like this
but maybe I'm just a sloppier coder). I see that the reviewer guidelines
say that I'm only allowed to "strongly encourage" you to implement tests so
you can do whatever you want but it doesn't take long to set up travis and
I hope you will consider it!

We have written a test for nearly every keyword argument and included them
in a tests directory under the package directory. We have made the test
function as a standalone script or one can run it using nosetests (nose is
still required in any case).

Otherwise, thanks for discussing and addressing all my comments and I think
that this looks much better. After these minor comments have been
addressed, I think we're good to go.

Thank you!
_____

On Sun, Sep 18, 2016 at 5:54 AM, Dan Foreman-Mackey <
notifications@github.com> wrote:

> @arfon <https://github.com/arfon>: is it possible to update the software
> version number and rebuild the paper?
>
> @SebastianBocquet <https://github.com/SebastianBocquet>: thanks for the
> responses – it's looking much better! I have a few minor comments remaining:
>
>    1.
>
>    The comment about "sigmas" has nothing to do with Gaussians. It has to
>    do with the relationship between the contours that you draw in 2D and the
>    quantiles that people tend to label in 1D. If you use 68% as a quantile in
>    1D then you shouldn't use the same in 2D because the contours will look
>    broader than they should!! I agree that there's no reason to choose a
>    specific number and encouraging people to say what the used is the best
>    practice but I'd be careful about the default value that you choose. I'd be
>    willing to make a large bet that when the vast majority of readers look at
>    plots like this they interpret the levels as 1 and 2 "sigma" even if you
>    say "68% contours" and they would be wrong. Also, "confidence" interval
>    won't, in general, be the correct term – if this is meant to display the
>    results of MCMC then these are actually credible regions
>    <https://en.wikipedia.org/wiki/Credible_interval>, for example. With
>    both of these points in mind, I'd update the terminology ("confidence
>    level" -> "contour level", "Gaussian" -> None) in the discussion and in the
>    docstring. (Sorry to harp on this so much but I've spent a lot of time
>    thinking about this and correcting people's misconceptions about these
>    types of plots over the years!)
>    2.
>
>    It's not actually true that adding version numbers to the setup.py
>    will not force upgrades (and it's actually generally considered bad
>    practice to put version numbers in the setup.py
>    <https://caremad.io/posts/2013/07/setup-vs-requirement/>). If tests
>    were included then it would be obvious which versions of the packages were
>    actually required 😉
>    3.
>
>    I totally disagree on the point about tests. The point of tests for
>    something like this is to confirm that the results of the sample plots are
>    the same for many different versions of numpy and matplotlib (trust me,
>    this is not a given – adding tests to corner.py found a few bugs like this
>    but maybe I'm just a sloppier coder). I see that the reviewer guidelines
>    say that I'm only allowed to "strongly encourage" you to implement tests so
>    you can do whatever you want but it doesn't take long to set up travis and
>    I hope you will consider it!
>
> Otherwise, thanks for discussing and addressing all my comments and I
> think that this looks much better. After these minor comments have been
> addressed, I think we're good to go.
>
