### 2016-10-03 #46 @dfm
Awesome work! This is getting really close.

I installed a clean Python 2.7 (I tried making a few small changes to support Python 3 but it might take some work!) conda environment with numpy, matplotlib, pandas, and nose. I also removed my .matplotlib directory so there aren't any rcParams set. With this installation, I ran the tests with the master branch of pyGTC and got [this list of failures](https://gist.github.com/dfm/d5d2097332439512bb897f36501746aa). I don't have scipy installed (this is a conda issue) so that error is okay (you should probably actually catch it and make it a pass if scipy isn't installed) but the others are unfortunate. I actually find that none of the figures are qualitatively similar and since I don't have _any_ rcParams set, I would argue that this isn't user error.

Compare bare.png:

![bare](https://cloud.githubusercontent.com/assets/350282/19046360/d7770214-8951-11e6-80f7-8ce9ac8b367a.png)

to bare-expected.png

![bare-expected](https://cloud.githubusercontent.com/assets/350282/19046367/ddb0482a-8951-11e6-9c38-a6c7534c6adf.png)
