### 2016-10-03 cc69649
Fixed tests in case no scipy and bumped version
