+++ pygtc/tests/test_plotGTC.py
@@ -1,12 +1,21 @@
+import warnings
 import numpy as np
 import matplotlib.pyplot as plt
 from matplotlib.testing.decorators import image_comparison
+import nose
 
 try:
     import pandas as pd
     HAS_PANDAS = True
 except ImportError:
     HAS_PANDAS = False
+
+try:
+    import scipy.ndimage
+    from scipy.stats import norm
+    HAS_SCIPY = True
+except ImportError:
+    HAS_SCIPY = False
 
 import pygtc
 
@@ -50,7 +59,8 @@ def test_img():
 #A test for (almost) every keyword argument
 @image_comparison(baseline_images=['bare'], extensions=['png'], savefig_kwarg=SFKWARGS)
 def test_GTC_bare():
-    pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2])
+    pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2],
+                    smoothingKernel = 0)
 
 @image_comparison(baseline_images=['pandas'], extensions=['png'], savefig_kwarg=SFKWARGS)
 def test_GTC_pandas():
@@ -59,35 +69,41 @@ def test_GTC_pandas():
     if HAS_PANDAS:
         samples1_pd = pd.DataFrame(SAMPLES_1, columns=namesNoTex)
         samples2_pd = pd.DataFrame(SAMPLES_2, columns=namesNoTex)
-    else:
-        samples1_pd = SAMPLES_1
-        samples2_pd = SAMPLES_2
 
-    pygtc.plotGTC(chains=[samples1_pd,samples2_pd])
+    else:
+        warnings.warn("Can't test pandas auto-name without pandas. Skipping test.", UserWarning)
+        raise nose.SkipTest
+
+    pygtc.plotGTC(chains=[samples1_pd,samples2_pd],
+                    smoothingKernel = 0)
 
 @image_comparison(baseline_images=['paramNames_noTex'], extensions=['png'], savefig_kwarg=SFKWARGS)
 def test_GTC_paramNames_noTex():
     namesNoTex = ['param name', 'B_labmda', 'C', 'lambda']
     pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2],
-                    paramNames = namesNoTex)
+                    paramNames = namesNoTex,
+                    smoothingKernel = 0)
 
 @image_comparison(baseline_images=['paramNames_withTex'], extensions=['png'], savefig_kwarg=SFKWARGS)
 def test_GTC_paramNames_withTex():
     namesWithTex = ['param name', '$B_\mathrm{\lambda}$', '$Q^a$', '$\\lambda$']
     pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2],
-                    paramNames = namesWithTex)
+                    paramNames = namesWithTex,
+                    smoothingKernel = 0)
 
 @image_comparison(baseline_images=['chainLabels_noTex'], extensions=['png'], savefig_kwarg=SFKWARGS)
 def test_GTC_chainLabels_noTex():
     chainLabelsNoTex = ['data1', 'data 2']
     pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2],
-                    chainLabels = chainLabelsNoTex)
+                    chainLabels = chainLabelsNoTex,
+                    smoothingKernel = 0)
 
 @image_comparison(baseline_images=['chainLabels_withTex'], extensions=['png'], savefig_kwarg=SFKWARGS)
 def test_GTC_chainLabels_withTex():
     chainLabelsWithTex = ['data1 $\lambda$', 'data 2']
     pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2],
[…]
