### 2016-10-04 #46 @SebastianBocquet
Your bare.png is not clearly not being smoothed, as opposed to
bare-expected.png, which is due to scipy not being installed, right? I
realize my answer was not very clear. Have you pulled the latest version of
pygtc? This really should fix the problem!

On Tue, Oct 4, 2016 at 1:53 PM, Dan Foreman-Mackey <notifications@github.com
> wrote:

> Nope! That isn't the problem. That was an error that should be caught (as
> I said above) but this doesn't address the actual problem that I brought up.
>
