### 2016-10-04 #46 @dfm
Oops. Sorry, I misread your comment. Looks good now. My last comment is that the tests don't close the figures. I'd change:

```
pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2], smoothingKernel = 0)
```

to:

```
fig = pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2], smoothingKernel=0)
plt.close(fig)
```

Otherwise this looks good. The only thing that I can't sign off on is the version number. I'm not sure if you need to do something or if @arfon does.
