### 2016-10-04 #46 @arfon
:zap: thanks @dfm.

@SebastianBocquet if you can make the final changes suggested then I think we're good to accept.

I'll need you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive. You might want to bump the release number to reflect the changes that have been made based on this review.
