### 2016-10-06 2f77a53
small tweak in test of prior plotting
