### 2016-10-06 #46 @SebastianBocquet
The test decorator handles cleaning up the figures. In fact, when trying
your suggestions, nosetests skips the tests.
Please confirm that we're still good to go, and I'll create a release for
the accepted version.

Thanks,
Sebastian

On Tue, Oct 4, 2016 at 5:35 PM, Dan Foreman-Mackey <notifications@github.com
> wrote:

> Oops. Sorry, I misread your comment. Looks good now. My last comment is
> that the tests don't close the figures. I'd change:
>
> pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2], smoothingKernel = 0)
>
> to:
>
> fig = pygtc.plotGTC(chains=[SAMPLES_1,SAMPLES_2], smoothingKernel=0)
> plt.close(fig)
>
> Otherwise this looks good. The only thing that I can't sign off on is the
> version number. I'm not sure if you need to do something or if @arfon
> <https://github.com/arfon> does.
>
