### 2016-10-06 #46 @dfm
Cool. I got a warning that too many figures were open when I ran the tests in Py3 but that must be because the figures don't get closed if the test fails – seems counter intuitive but that is all I can think of.

Looks good to me!
