### 2016-10-06 #46 @arfon
@dfm - can you confirm we're good to accept at this point?

@SebastianBocquet if @dfm confirms, the next step is to make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
