### 2016-10-07 #46 @SebastianBocquet
And here it is: DOI: 10.5281/zenodo.159225
<https://doi.org/10.5281/zenodo.159225>

Thanks,
Sebastian

On Wed, Oct 5, 2016 at 9:39 PM, Dan Foreman-Mackey <notifications@github.com
> wrote:

> 👍 good to go!
>
