### 2016-10-08 #46 @arfon
@dfm - many thanks for the thorough review.

@SebastianBocquet - your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00046 . 🎉 🚀 💥
